+++ gateway-api-integrator/tests/integration/test_actions.py
@@ -11,11 +11,22 @@ from juju.application import Application
 @pytest.mark.abort_on_fail
 async def test_get_certificate_action(
     configured_application_with_tls: Application,
+    ingress_requirer_application: Application,
 ):
     """Deploy the charm with valid config and tls integration.
 
     Assert on valid output of get-certificate.
     """
+    await configured_application_with_tls.model.add_relation(
+        configured_application_with_tls.name,
+        f"{ingress_requirer_application.name}:ingress",
+    )
+    await configured_application_with_tls.model.wait_for_idle(
+        apps=[configured_application_with_tls.name, ingress_requirer_application.name],
+        idle_period=30,
+        status="active",
+    )
+
     action = await configured_application_with_tls.units[0].run_action(
         "get-certificate", hostname=TEST_EXTERNAL_HOSTNAME_CONFIG
     )