+++ gateway-api-integrator/tests/unit/test_certificates_integration.py
@@ -6,6 +6,7 @@
 import pytest
 from ops.testing import Harness
 
+from state.charm_state import CharmState, ProxyMode
 from state.tls import TLSInformation, TlsIntegrationMissingError
 
 
@@ -17,9 +18,16 @@ def test_tls_information_integration_missing(harness: Harness):
     assert: TLSIntegrationMissingError is raised.
     """
     harness.begin()
+    charm_state = CharmState(
+        gateway_class_name="cilium",
+        enforce_https=True,
+        proxy_mode=ProxyMode.INGRESS,
+        requires_ip_certificate=False,
+        hostnames={"example.com"},
+    )
     with pytest.raises(TlsIntegrationMissingError):
         TLSInformation.from_charm(
             harness.charm,
-            "example.com",
+            charm_state.hostnames,
             harness.charm.certificates,
         )